replaced BaseController with Controller
