+++ .gitignore
@@ -5,7 +5,6 @@
 /.idea
 /.vagrant
 /.env
-/.env.example
 Homestead.json
 Homestead.yaml
 npm-debug.log
